major development update.
* removed all old code.

the old code was playing hell on `go mod`, so I finally just removed it
to simplify the dependency tree and make things much more sane.

* extracted gravel.

gravel can now be found at https://github.com/18f/gravel as it's own
project with it's own lifecycle. now this project just consumes it.

* got rid of some of the fakes.

I realised I needed mocks, not fakes, so some fakes are now mocks.

* fixed modules.

the dependency tree was hell. now it's sane.

* passing tests.

now the domain broke can pass the tests in the broker suite, which is
great! took a bit to get there, but now we're here. so that's nice.

Signed-off-by: Mike Lloyd <mike.lloyd@gsa.gov>